+++ src/context/AppContext.js
@@ -6,19 +6,19 @@ import { useTranslation } from 'react-i18next';
 import { ReferenceContext } from './ReferenceContext';
 import { getResources, getBookList, checkLSVal, getLayoutType } from '../helper';
 import {
-  server,
   defaultTplBible,
   defaultTplOBS,
   languages,
   bibleList,
+  server,
 } from '../config/base';
+
 export const AppContext = React.createContext();
 
 const _currentLanguage = checkLSVal('i18nextLng', languages[0]);
 const _resourcesApp = checkLSVal('resourcesApp', [], 'object');
 const _startDialog = checkLSVal('startDialog', true, 'boolean');
 const _fontSize = parseInt(localStorage.getItem('fontSize'));
-const config = { server };
 
 export function AppContextProvider({ children }) {
   const {
@@ -41,7 +41,11 @@ export function AppContextProvider({ children }) {
   );
 
   const [breakpoint, setBreakpoint] = useState({ name: 'lg', cols: 12 });
-
+  /** TODO Create ResourceContext
+   * 1. Get information about resources ( like available bookId) from /Chapter  - content.resources.project.
+   * 2. Put all states about resources in ResourceContext.
+   * 3. Maybe make availableBookList in ResourceContext
+   */
   const [resourcesApp, setResourcesApp] = useState(_resourcesApp);
   const _resourceLinks = getResources(appConfig, resourcesApp);
   const [resourceLinks, setResourceLinks] = useState(_resourceLinks);
@@ -54,7 +58,7 @@ export function AppContextProvider({ children }) {
   const [loadIntro, setLoadIntro] = useState(false);
   const [openStartDialog, setOpenStartDialog] = useState(_startDialog);
   const [openMainMenu, setOpenMainMenu] = useState(false);
-
+  const config = { server };
   const { t } = useTranslation();
 
   useEffect(() => {